documentation plus minor changes
removed -march=native from bindings to allow distribution
replaces SE_MAX_STNAME with AS_MAXCH
corrected integer type mismatches with the C library

+++ src/functions/d2l.cpp
@@ -10,6 +10,6 @@ Napi::Value sweph_d2l(const Napi::CallbackInfo& info) {
 	if(!sweph_type_check(args, info)) {
 		return env.Null();
 	}
-	long val = swe_d2l(info[0].As<Napi::Number>().DoubleValue());
+	int32 val = swe_d2l(info[0].As<Napi::Number>().DoubleValue());
 	return Napi::Number::New(env, val);
 }
